worker/requirements.txt
celery[redis]==5.5.*
supabase==2.28.*
httpx==0.28.*
python-dotenv==1.1.*
pydantic-settings==2.9.*
google-genai==1.14.*
